+++ config/BUILD.bazel
@@ -9,9 +9,8 @@ go_library(
     importpath = "github.com/bazelbuild/bazel-gazelle/config",
     visibility = ["//visibility:public"],
     deps = [
-        "//internal/module",
-        "//internal/wspace",
-        "//rule",
+        "//v2/config",
+        "//v2/rule",
     ],
 )
 
@@ -19,7 +18,7 @@ go_test(
     name = "config_test",
     srcs = ["config_test.go"],
     embed = [":config"],
-    deps = ["//rule"],
+    deps = ["//v2/rule"],
 )
 
 filegroup(
