+++ v2/config/BUILD.bazel
@@ -1,0 +1,23 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "config",
+    srcs = ["config.go"],
+    importpath = "github.com/bazel-contrib/bazel-gazelle/v2/config",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//v2/internal/module",
+        "//v2/internal/wspace",
+        "//v2/rule",
+    ],
+)
+
+filegroup(
+    name = "all_files",
+    testonly = True,
+    srcs = [
+        "BUILD.bazel",
+        "config.go",
+    ],
+    visibility = ["//visibility:public"],
+)
